add updateBlog and deleteBlog

+++ src/main/java/com/oeps/onlineeventplanningsystem/repositories/BlogRepo.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface BlogRepo extends JpaRepository<Blog, Long> {
+public interface BlogRepo extends JpaRepository<Blog, Integer> {
 
     Optional<Blog> findBlogByBlogTitle(String blogTitle);
